requirements.txt
asgiref==3.11.0
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==5.2.10
django-filter==25.2
djangorestframework==3.14.0
idna==3.11
isodate==0.7.2
lxml==6.0.2
platformdirs==4.5.1
pytz==2025.2
requests==2.31.0
requests-file==3.0.1
requests-toolbelt==1.0.0
spyne==2.14.0
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
zeep==4.2.1
pymysql==1.1.1
mysql-connector-python==9.5.0
cryptography==42.0.5